cavern_stronghold: event cleanup
Most locations are now identifiers in the map itself, and events were refactored
accordingly.

Replaced `moveto` misuse with the more-appropriate enter_hex and exit_hex.

Added the beginning of event handling for the first encounter.

The GM now has spaces to move to to trigger manual story events as well.